Fixed move bugs / added pivot ability
Fixed some problems with the move_piece code.

Added the ability to pivot a piece based on the 0 based pivot position. This
means that a ship at ((4,1), (4,2), (4,3)) with a pivot value of 1 will flip on
position (4,2) so that the new coordinates will be ((3,2), (4,2), (5,2)).

Some unit tests have been created but more are needed.